Final 2020 - Day 10 Solution (with notes!)
Our final 2020 Day 10 solution is maximum python readability as
well as maximum algorithmic efficiency.

We know that path lengths spike in possibility based on the chains
with consecutive runs (gaps of 1 instead of gaps of 3). We also
know (by analyzing the data) that the maximum "chain" size is 5.
So we can that knowledge to mulitply out the possibilities.

We can also calculate the run as we go.

We take advantage of some special characteristics of enumerate()
to simplify calling and logic.