Cleanup 2021 - Day 07 - Fix Part 2
Part 2 in my original code had an unfortuante error of data. The
logical average (458) was not actually the right target to go for.
Instead, it was 457 that gave me trouble. To fix this for any
input, I wrote some code to check above and below the average,
so that whatever it was, it would still give me the right answer.