Add namedtuple() and prod/test print at EOF
I've now run into issues twice where I've run a puzzle with the
test data and submitted it. The code was correct! The data was not.
So now, we have some extra checks at the end of every file to
properly notify the end user (aka, me) as to what version of the
data is running. There's the fancy test/prod bit at the top, but
that doesn't help when the output garbos the screen.

I've also gone and added in namedtuple() because it's something I
use quite often in these puzzles.