Significantly speed up determine_shortest()
My first implementation of determine_shortest() used a very slow,
naive approach to finding the next node to look at. I simply
re-iterated over the entire array for each node.

But that's a _huge_ problem as we get a larger and larger input
size.

So my second attempt uses a smarter approach, only going through
the edge nodes of what we want to check -- a significantly smaller
data set.